chore: implement for any codable
