+++ Sources/Notifly/notifly-ios-sdk/notifly-ios-sdk/SourceCodes/NotiflyUtil/AppHelper.swift
@@ -1,6 +1,7 @@
 import Foundation
 import Security
 import UIKit
+
 class AppHelper {
     static func getNotiflyDeviceID() -> String? {
         guard let deviceID = AppHelper.getDeviceID() else {
@@ -55,33 +56,6 @@ class AppHelper {
         return Bundle.main.bundleIdentifier
     }
 
-    static func makeJsonCodable(_ jsonData: [String: Any]?) -> [String: AnyCodable]? {
-        guard let jsonData = jsonData else { return nil }
-        return jsonData.mapValues { value in
-            if let array = value as? [Any?] {
-                return AnyCodable(array.compactMap { element in AppHelper.toCodableValue(element) })
-            } else if let dictionary = value as? [String: Any] {
-                return AnyCodable(makeJsonCodable(dictionary))
-            }
-            return AppHelper.toCodableValue(value)
-        }
-    }
-
-    static func toCodableValue(_ value: Any?) -> AnyCodable {
-        if let str = value as? String {
-            return AnyCodable(str)
-        } else if let int = value as? Int {
-            return AnyCodable(int)
-        } else if let double = value as? Double {
-            return AnyCodable(double)
-        } else if let float = value as? Float {
-            return AnyCodable(float)
-        } else if let bool = value as? Bool {
-            return AnyCodable(bool)
-        } else {
-            return AnyCodable(value)
-        }
-    }
 }
 
 private func saveUniqueIdToKeychain(deviceID: String) -> Bool {
@@ -90,7 +64,7 @@ private func saveUniqueIdToKeychain(deviceID: String) -> Bool {
         let query: [String: Any] = [
             kSecClass as String: kSecClassGenericPassword,
             kSecAttrAccount as String: uniqueIdKey,
-            kSecValueData as String: deviceID.data(using: .utf8)!,
+            kSecValueData as String: deviceID.data(using: .utf8)!
         ]
 
         let status = SecItemAdd(query as CFDictionary, nil)
@@ -108,7 +82,7 @@ private func retrieveUniqueIdFromKeychain() -> String? {
             kSecClass as String: kSecClassGenericPassword,
             kSecAttrAccount as String: uniqueIdKey,
             kSecReturnData as String: true,
-            kSecMatchLimit as String: kSecMatchLimitOne,
+            kSecMatchLimit as String: kSecMatchLimitOne
         ]
         var result: AnyObject?
         let status = SecItemCopyMatching(query as CFDictionary, &result)
