+++ Sources/Notifly/notifly-ios-sdk/notifly-ios-sdk/SourceCodes/NotiflyUtil/TrackingEvent.swift
@@ -76,57 +76,6 @@ struct TrackingData: Codable {
     let event_params: [String: AnyCodable]?
 }
 
-struct AnyCodable: Codable {
-    private let value: Any
-
-    init(_ value: Any) {
-        self.value = value
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var container = encoder.singleValueContainer()
-
-        if let encodable = value as? Encodable {
-            try encodable.encode(to: encoder)
-        } else {
-            try container.encodeNil()
-        }
-    }
-
-    func getValue() -> Any {
-        return self.value
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-
-        if let intValue = try? container.decode(Int.self) {
-            value = intValue
-        } else if let stringValue = try? container.decode(String.self) {
-            value = stringValue
-        } else if let boolValue = try? container.decode(Bool.self) {
-            value = boolValue
-        } else if let doubleValue = try? container.decode(Double.self) {
-            value = doubleValue
-        } else if let arrayValue = try? container.decode([AnyCodable].self) {
-            value = arrayValue
-        } else if let floatValue = try? container.decode(Float.self) {
-            value = floatValue
-        } else if let dictionaryValue = try? container.decode([String: AnyCodable].self) {
-            value = dictionaryValue
-        } else if let arrayValue = try? container.decode([AnyCodable?].self) {
-            value = arrayValue.compactMap { $0?.value }
-        } else if container.decodeNil() {
-            value = ()
-        } else {
-            throw DecodingError.dataCorruptedError(
-                in: container,
-                debugDescription: "Unsupported type"
-            )
-        }
-    }
-}
-
 public enum SdkWrapperType: String {
     case react_native
     case flutter
